Integrate smtmatchers.cpp into smt.cpp
