+++ CMakeLists.txt
@@ -2,10 +2,10 @@ cmake_minimum_required(VERSION 3.15.0)
 project(mlir-tv VERSION 0.1.0)
 set (CMAKE_CXX_STANDARD 17)
 
-set(MLIR_DIR CACHE PATH "MLIR installation top-level directory")
-set(Z3_DIR CACHE PATH "Z3 installation top-level directory")
-set(CVC5_DIR CACHE PATH "CVC5 installation top-level directory")
-option(USE_LIBC "Use libc++ in case the MLIR (and CVC5) is linked against libc++")
+set(MLIR_DIR /opt/llvm-13-libc CACHE PATH "MLIR installation top-level directory")
+set(Z3_DIR /opt/z3-inst CACHE PATH "Z3 installation top-level directory")
+set(CVC5_DIR /opt/cvc5 CACHE PATH "CVC5 installation top-level directory")
+option(USE_LIBC "Use libc++ in case the MLIR is linked against libc++" ON)
 
 set(MLIR_INC_DIR "${MLIR_DIR}/include")
 set(MLIR_LIB_DIR "${MLIR_DIR}/lib")
@@ -22,7 +22,6 @@ add_library(${PROJECT_LIB} STATIC
     src/memory.cpp
     src/print.cpp
     src/smt.cpp
-    src/smtmatchers.cpp
     src/state.cpp
     src/value.cpp
     src/vcgen.cpp)
@@ -74,6 +73,17 @@ if(CVC5_DIR)
         target_link_directories(${PROJECT_LIB} PUBLIC ${CVC5_LIB_DIR})
         target_link_libraries(${PROJECT_LIB} PUBLIC cvc5)
         target_compile_definitions(${PROJECT_LIB} PUBLIC SOLVER_CVC5)
+    endif()
+endif()
+
+# Check for CVC5 installation
+if(CVC5_DIR)
+    if(NOT EXISTS ${CVC5_INC_DIR} OR NOT EXISTS ${CVC5_LIB_DIR})
+        message(FATAL_ERROR "the provided path doesn't seem to be a valid CVC5 directory")
+    else()
+        target_include_directories(${PROJECT_LIB} PUBLIC ${CVC5_INC_DIR})
+        target_link_directories(${PROJECT_LIB} PUBLIC ${CVC5_LIB_DIR})
+        target_link_libraries(${PROJECT_LIB} PUBLIC cvc5)
     endif()
 endif()
