+++ src/smt.h
@@ -19,21 +19,6 @@
 #endif
 
 namespace smt {
-
-class Context {
-public:
-  std::optional<z3::context> z3;
-
-  Context() {}
-  Context(bool use_z3) {
-    if (use_z3) {
-      this->z3.emplace();
-    }
-  }
-};
-
-extern Context sctx;
-
 class Expr;
 class FnDecl;
 class Model;
@@ -135,7 +120,6 @@ public:
   static Expr mkAddNoOverflow(const Expr &a, const Expr &b, bool is_signed);
 
 
-  friend Context;
   friend FnDecl;
   friend Model;
   friend Solver;
